requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==1.17.1
charset-normalizer==3.4.3
cssselect2==0.8.0
Django==4.2
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fonttools==4.59.0
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.3
setuptools==80.9.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
zopfli==0.2.3.post1